+++ app/internal/localauth/login.go
@@ -135,12 +135,6 @@ func serveLoginSubmit(w http.ResponseWriter, r *http.Request) error {
 	// Authenticate future requests.
 	ctx = sourcegraph.WithCredentials(ctx, oauth2.StaticTokenSource(&oauth2.Token{TokenType: "Bearer", AccessToken: tok.AccessToken}))
 
-	// Fetch username.
-	authInfo, err := cl.Auth.Identify(ctx, &pbtypes.Void{})
-	if err != nil {
-		return err
-	}
-
 	// Authenticate as newly created user.
 	if err := appauth.WriteSessionCookie(w, appauth.Session{AccessToken: tok.AccessToken}); err != nil {
 		return err
@@ -153,7 +147,7 @@ func serveLoginSubmit(w http.ResponseWriter, r *http.Request) error {
 		return err
 	}
 	if returnTo == "" {
-		returnTo = router.Rel.URLToUser(authInfo.Login).String()
+		returnTo = "/"
 	}
 
 	http.Redirect(w, r, returnTo, http.StatusSeeOther)