+++ app/router/router.go
@@ -7,7 +7,6 @@ import (
 	"os"
 
 	"github.com/sourcegraph/mux"
-	"src.sourcegraph.com/sourcegraph/auth/authutil"
 	"src.sourcegraph.com/sourcegraph/conf/feature"
 	gitrouter "src.sourcegraph.com/sourcegraph/gitserver/router"
 	"src.sourcegraph.com/sourcegraph/go-sourcegraph/routevar"
@@ -36,13 +35,10 @@ const (
 	SitemapIndex = "sitemap-index"
 	RepoSitemap  = "repo.sitemap"
 
-	User                      = "person"
 	UserSettingsEmails        = "person.settings.emails"
 	UserSettingsProfile       = "person.settings.profile"
 	UserSettingsProfileAvatar = "person.settings.profile.avatar"
 	UserSettingsKeys          = "person.settings.keys"
-	UserOrgs                  = "person.orgs"
-	OrgMembers                = "org.members"
 
 	Repo             = "repo"
 	RepoBadge        = "repo.badge"
@@ -165,11 +161,6 @@ func New(base *mux.Router) *Router {
 	user.Path("/.settings/profile/avatar").Methods("POST").Name(UserSettingsProfileAvatar)
 	user.Path("/.settings/emails").Methods("GET").Name(UserSettingsEmails)
 	user.Path("/.settings/keys").Methods("GET", "POST").Name(UserSettingsKeys)
-	if !authutil.ActiveFlags.DisableUserProfiles {
-		base.Path(userPath).Methods("GET").Name(User)
-		user.Path("/.orgs").Methods("GET").Name(UserOrgs)
-		user.Path("/.members").Methods("GET").Name(OrgMembers)
-	}
 
 	// attach git transport endpoints
 	gitrouter.New(base)