+++ app/script/components/DeltaAuthor.js
@@ -10,9 +10,9 @@ var DeltaAuthor = React.createClass({
 		var da = this.props.deltaAuthor;
 		return (
 			<li className="list-group-item">
-			<a className="pull-left" href={da.Login && router.personURL(da.Login)}><img src={da.AvatarURL} className="media-object avatar img-rounded" width="40" /></a>
+			<img src={da.AvatarURL} className="pull-left media-object avatar img-rounded" width="40" />
 			<div className="media-body">
-			<h3><span className="affected"><a className="pull-left" href={da.Login && router.personURL(da.Login)}>{Person.label(da)}</a></span> &nbsp;contributed to:</h3>
+			<h3><span className="affected pull-left">{Person.label(da)}</span> &nbsp;contributed to:</h3>
 			<ul className="list-group defs">
 				{da.Defs.map(function(def) {
 					var df = def.FmtStrings;